Fix bugs
1. Bug in chrome browser. Fix token authentication.

2. Bug in DB driver. Fix QueryBuilder.

+++ controller/PostController.php
@@ -263,7 +263,12 @@ class PostController
 
 		$image = App::get('public_path') . '/storage/images/' . $post['image'];
 
-		unlink($image);
+
+		if ($image) {
+
+			unlink($image);
+
+		}
